+++ solidity/choices/FutureAsyncCondChoice.sol
@@ -20,7 +20,7 @@ contract FutureAsyncCondChoice is AbstractAsyncChoice {
     super.activateEvent(index);
   }
 
-  function oracleCallback(uint256 correlation, bytes calldata) external override {
+  function oracleCallback(uint16 correlation, bytes calldata) external override {
     // Do nothing if we have already finished
     if (winner >= 0) {
       return;
